feat: Add FinishIResult::finish to drop remaining input
It seems like the need for this is common enough to be worth having,
especially to drive correct behavior as this change highlights several
examples that ignored remaining input even when that isn't correct
(like in a `FromStr`).

This does add extra bounds for `FinishIResult`.  I had considered making
a separate extension trait but the likelihood it would be needed is low.
If anything, I could see adding an `InputIsEmpty` trait and moving it to
that as that would be more universal (could work on some iterators).  If
we really need it, we can split the traits (create a
`FinishErrIResult`).  Let's see what feedback we get.