feat: Add Parser::dbg_err
This is a translation of `dbg_dmp`.

The name changed because we are only printing errors, we might want to
print something else in the future

Random thoughts on the name
- `debug` is out because that implies a debug build (`debug_assert`)
- `dbg` mirrors `dbg!` for dumping content to output
- `_err` is a common suffix (`unwrap_err`)
- We want this easily discoverably to audit to prevent going into
  production.  People likely will audit for `dbg` so having a similar
  name will likely get this caught as well.

I had also considered `dump_err` which I favored until I considered code
audits.

As for functionality, this adds
- Support for any input that supports `AsBytes`, instead of just `&[u8]`
- Allows any displayable context
- Print to stderr, instead of stdout