fix(error)!: Deprecate `from_char` in favor of ContextError
`from_char` is a bit specialized, only working with one combinator that
can only work with `&str` and not `&[u8]` (which the latter is usually
needed in high performance cases).

Instead, users can add all sorts of context
to an error, either with multiple implementations or with an
implementation that accepts an `enum` now that `ContextError` is generic.

This isn't automatic but the old solution was barely there for automatic
behavior and automatic behavior has a ton of negative side effects
(performance, not quite right messages)