fix(parser)!: Allow literals to be parsers of partial inputs
If you used `Partial`, you couldn't use `"hello"` as a tag but had to do
`tag("hello")` because we needed to constrain the generic parameter.
Now that we've removed it, literals can be used as desired.

The downside is that for `FinishIResult`, this shifts an error from
compile time to runtime.  At least with `is_partial_support`, the user
is more likely to see it than they would have with the original `nom`
code.

BREAKING CHANGE: `StreamIsPartial` no longer takes a `const PARTIAL:
bool` generic parameter.  This was replaced with `is_partial_support`
function.

Fixes #79