docs: Split long-form docs into reference or cookbook
The scattered long-form made it harder to read the documentation and
were less discoverable.  This splits the content between the API
reference directly on the relevant items and on the cookbook.  This
ended up removing a lot of duplication.  For example, we talked about
the `ErrMode` variants several times over.

This left the tutorial in a mess but a followup with focus just on that.