fix(char)!: Allow empty `escape`
rust-bakery/nom#958 required at least one non-escaped value but only for
`complete` and  not for `streaming`.  There was also no explanation of
use case or discussion.  Generally, escaping is used for strings and
strings are generally allowed to be empty.

There is the question of whether it should be renamed to `escape0`.  I'm
setting that aside for  now.