perf: Undo performance regression from 0.4
Updates #223. This brings the peformance of the space parsers back to
pre-0.4.0 levels.  The root cause is not clear and the performance
difference shown does not align with the `contains_token` benchmark.

This only fixes half of the performance regression from
https://github.com/winnow-rs/winnow/pull/210 though. There are still
some other changes that negatively affect performance which I didn't
figure out yet.