docs: Add `hcl-edit` to the list of parsers using `winnow`
Our recent discussion in https://github.com/winnow-rs/winnow/issues/223
pushed me to finally make the initial release of `hcl-edit` which I've
been holding of for too long already since I got carried away with
missing details.

The crate is far from complete, but the parser implementation is
functional and actually quite fast.

So I figured I'll add it to the list of parsers for others to explore
real-world uses of winnow.

I hope this does not scare away people from winnow since HCL is quite a
big and messy piece to parse, given the three sub-languages it has and
some other quirks like the absense of keywords (for example, `true` can
be an identifier or a boolean, depending on the context) ;).