fix(trace)!: Force parsers to be nice
This hack was added in #156 but not seeing what caused it.  In general,
we should probably instead fix these parsers.

The problem with this hack is its isolated to `trace`, so other parts of
`winnow` may break if the code changes slightly.  Instead, we should
shoot for general correctness.  The problem with moving the hack into
`Offset` impls is that that is now in the inner loop, slowing things
down.