fix(error)!: Rename ParseError to ParserError
`ParseError` is a holdover from when it was being returned by
`Parser::parse` but now it is being returned by `Parser::parse_next` or
`Parser::parse_peek`.

The new name is to convey this is what the parser is returning.  This is
opening room for a `ParseError` which will be a concrete type that wraps
`ParserError` when returned by `Parser::parse`.