fix(parser)!: Switch char from one_of to literal
Originally, this was going to be done for performance reasons but when
comparing this to the previous commit, `tag(char)` / `char` slowed down
dramatically in the `next_slice` benchmark.

However, `json` benchmark was unchanged.  This also (slightly)
simplifies traces, is more direct in what it says, and does still speed
things up for UTF-8-exclusive `char`s, so going forward with it.

Fixes #418

BREAKING CHANGE: `char` parers now always return `char` as the output
and now require `I: Compare<char>`