Improve repeat_till doc's map then take advice
I tried to do `.map(|()| ())` as currently recommended when trying to take a series of tokens. This gave me a compiler error saying that I needed to put a tuple instead for the args in the map. Switching to `.map(|((), _)| ())` fixed my problem, accumulating the Accumulator into nothing and discarding the result of the combinator that repeat_till took.

This was a stumbling block for me, and I thought this tweak might make things easier for the next person.