refactor(parser): Manually inline Parser struct new's
This reduces the boilerplate for writing an opaque `impl Parser`.  In
this case, 240 lines.

I originally did the `new`s as I was thinking I might want to use it for
other purposes and I was less comfortable with writing all of that
PhantomData at the time (which only grew since with #222).