fix(token)!: Dont require SliceLen for `literal`
This makes it easier to use `literal` with more types.
I also expect this doesn't lose out on much as `Needed` for a literal is
not likely to report the actual number of needed tokens but a small
amount for the next batch, likely too small to be worth trying to fetch
an exact amount.