feat(stream): Allow unchecked slicing
There are times when the caller knows that the `offset` is good and can
speed things up by bypassing any checks explicitly.

For now, we're not takin advantage of this inside of `winnow`.
I want to be sure its safety to rely on all of the trait behaviors
before doing so.
When we do, we might want to put it behind an `unsafe` feature so people
can opt-in to performance with a risk.

I made these new functions inherent for more than compatibility reasons
but to allow people to have `#[forbid(unsafe)]` in there code.