feat: Dynamic add context to Context
This MR adds a new `std`-only feature to `Context`, which allows users
to dynamically append error context information based on runtime
variables.

This was previously not possible as the only way to dynamically append
information was to add another `.context()` call, which wrapped the
whole parser in a new type, thereby changing the type signature.

This change adds alters the `Context` in a backwards-compatible way by
changing the `Context::context` field to a `Vec<C>` and adding a
`add_context` function to add context to said vector.