Fix overflows for BStr PartialEq and PartialOrd
Using the BStr PartialEq or PartialOrd impls with &str or &[u8] results
in overflows due to the lack of well defined base cases for the impls.

This commit fixes the issue by removing the final layer of derived impls
and manually defining the base cases to ensure the recursion stops.

Tests are also added to help catch this issue in the future.