Implement `<&str as Stream>::next_token()` without panics
This PR makes `<&str as Stream>::next_token()` use `Chars` both for
finding the first character, and the remaining string. While the
compiler could not tell that the previous implementation cannot panic,
it can for the new one. Also, the offset does not have to be calculated
twice. All-in-all the method has now about 1/4 fewer instructions.

* Old: <https://godbolt.org/z/fTbqch3v8>
* New: <https://godbolt.org/z/MPx7PMMaW>