feat(comb): Add an operator precedence parser
This commit adds the `expression` module, which implements an API for
operator precedence parsing. It specifically leverages a Pratt parsing
algorithm.

The core of the API is the `Expression` struct, which implements a
builder pattern for configuring the Pratt parser. You create it using
the `expression()` method.

This is based on the work by @39555 in PR #614, with some fixes and
style changes by @ssmendon.

Closes #131

Co-authored-by: Igor <ig_@tutamail.com>